Rename icon components folder
