+++ src/lib/components/buttons/CollapseButton.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import RotatingIcon from '../icon/RotatingIcon.svelte';
-	import { IconSvg } from '../icon/icon-svg';
-	import { defaultRotatingIconData, rotateIcon } from '../icon/rotating-icon';
+	import RotatingIcon from '../icons/RotatingIcon.svelte';
+	import { IconSvg } from '../icons/icon-svg';
+	import { defaultRotatingIconData, rotateIcon } from '../icons/rotating-icon';
 
 	// Collapsed by default
 	export let icon: IconSvg.ChevronUp | IconSvg.ChevronDown = IconSvg.ChevronDown;
