+++ src/lib/components/icons/icon-svg.ts
