+++ src/lib/components/icons/rotating-icon.ts
